+++ src/components/myPage/tabs/resumeTab/ResumeTab.tsx
@@ -1,8 +1,8 @@
 import { styled } from "styled-components";
-import ResumeExperienceComp from "./ResumeExperienceComp";
-import ResumeProfileIntroComp from "./ResumeProfileIntroComp";
 import React from "react";
 import { S } from "./Resume.styles";
+import ResumeProfileIntroComp from "./profileIntro/ResumeProfileIntroComp";
+import ResumeExperienceComp from "./experience/ResumeExperienceComp";
 
 const ResumeTab = () => {
   return (
